+++ dnsttEx-client/main_tunnel.go
@@ -30,6 +30,7 @@ func runTunnel(
 	scanChecks int,
 	clientMTUFlag int,
 	sendParallel int,
+	tunnelMode string,
 ) error {
 	endpoints := make([]*poolEndpoint, 0, len(specs))
 	for _, spec := range specs {
@@ -142,5 +143,8 @@ func runTunnel(
 	}
 
 	pconn = NewDNSPacketConn(pconn, remoteAddr, domain, effectiveMaxResponse, effectiveMaxRequest)
+	if tunnelMode == "socks" {
+		return runSocks(pubkey, domain, localAddr, remoteAddr, pconn)
+	}
 	return run(pubkey, domain, localAddr, remoteAddr, pconn)
 }